# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Frontier Road API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: bounties
    description: Bounty board operations
  - name: residents
    description: Resident hub operations
  - name: treasury
    description: Treasury operations
  - name: openai
    description: OpenAI AI chat, image, and audio operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /bounties:
    get:
      operationId: listBounties
      tags: [bounties]
      summary: List all bounties
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, claimed, completed, cancelled]
      responses:
        "200":
          description: List of bounties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bounty"
    post:
      operationId: createBounty
      tags: [bounties]
      summary: Create a new bounty
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBountyBody"
      responses:
        "201":
          description: Created bounty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bounty"

  /bounties/{id}:
    get:
      operationId: getBounty
      tags: [bounties]
      summary: Get a single bounty
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bounty details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bounty"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /bounties/{id}/claim:
    post:
      operationId: claimBounty
      tags: [bounties]
      summary: Claim a bounty
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimBountyBody"
      responses:
        "200":
          description: Claimed bounty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bounty"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Already claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /bounties/{id}/complete:
    post:
      operationId: completeBounty
      tags: [bounties]
      summary: Submit proof and complete a bounty
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteBountyBody"
      responses:
        "200":
          description: Completed bounty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bounty"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /bounties/{id}/cancel:
    post:
      operationId: cancelBounty
      tags: [bounties]
      summary: Cancel a bounty and release escrow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled bounty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bounty"

  /residents:
    get:
      operationId: listResidents
      tags: [residents]
      summary: List all residents
      parameters:
        - name: skill
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of residents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resident"
    post:
      operationId: createResident
      tags: [residents]
      summary: Register a new resident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResidentBody"
      responses:
        "201":
          description: Created resident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resident"

  /residents/{id}:
    get:
      operationId: getResident
      tags: [residents]
      summary: Get a single resident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resident details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resident"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateResident
      tags: [residents]
      summary: Update a resident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResidentBody"
      responses:
        "200":
          description: Updated resident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resident"

  /treasury:
    get:
      operationId: getTreasury
      tags: [treasury]
      summary: Get treasury overview
      responses:
        "200":
          description: Treasury state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreasuryOverview"

  /treasury/transactions:
    get:
      operationId: listTransactions
      tags: [treasury]
      summary: List treasury transactions
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"

  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpenaiConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"

  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"

  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenaiMessageBody"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}

  /openai/conversations/{id}/voice-messages:
    post:
      operationId: sendOpenaiVoiceMessage
      tags: [openai]
      summary: Send audio and receive a streaming voice response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenaiVoiceMessageBody"
      responses:
        "200":
          description: SSE stream of transcript and audio chunks
          content:
            text/event-stream: {}

  /openai/generate-image:
    post:
      operationId: generateOpenaiImage
      tags: [openai]
      summary: Generate an image from a text prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateOpenaiImageBody"
      responses:
        "200":
          description: Generated image data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateOpenaiImageResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Bounty:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        rewardAmount:
          type: number
        rewardToken:
          type: string
        status:
          type: string
          enum: [open, claimed, completed, cancelled]
        category:
          type: string
        creatorWallet:
          type: string
        claimerWallet:
          type: string
          nullable: true
        proofOfWork:
          type: string
          nullable: true
        escrowTxSignature:
          type: string
          nullable: true
        completionTxSignature:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - rewardAmount
        - rewardToken
        - status
        - category
        - creatorWallet
        - createdAt
        - updatedAt

    CreateBountyBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        rewardAmount:
          type: number
        rewardToken:
          type: string
          default: USDC
        category:
          type: string
        creatorWallet:
          type: string
        escrowTxSignature:
          type: string
      required:
        - title
        - description
        - rewardAmount
        - category
        - creatorWallet

    ClaimBountyBody:
      type: object
      properties:
        claimerWallet:
          type: string
      required:
        - claimerWallet

    CompleteBountyBody:
      type: object
      properties:
        proofOfWork:
          type: string
        completionTxSignature:
          type: string
      required:
        - proofOfWork

    Resident:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        walletAddress:
          type: string
          nullable: true
        avatar:
          type: string
          nullable: true
        skills:
          type: array
          items:
            type: string
        floor:
          type: integer
          nullable: true
        status:
          type: string
          enum: [online, offline, busy]
        bio:
          type: string
          nullable: true
        bountiesCompleted:
          type: integer
        bountiesCreated:
          type: integer
        totalEarned:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - skills
        - status
        - bountiesCompleted
        - bountiesCreated
        - totalEarned
        - createdAt

    CreateResidentBody:
      type: object
      properties:
        name:
          type: string
        walletAddress:
          type: string
        avatar:
          type: string
        skills:
          type: array
          items:
            type: string
        floor:
          type: integer
        bio:
          type: string
      required:
        - name
        - skills

    UpdateResidentBody:
      type: object
      properties:
        name:
          type: string
        walletAddress:
          type: string
        avatar:
          type: string
        skills:
          type: array
          items:
            type: string
        floor:
          type: integer
        status:
          type: string
          enum: [online, offline, busy]
        bio:
          type: string

    TreasuryOverview:
      type: object
      properties:
        totalBalance:
          type: number
        pendingEscrow:
          type: number
        totalPaidOut:
          type: number
        activeBounties:
          type: integer
        completedBounties:
          type: integer
      required:
        - totalBalance
        - pendingEscrow
        - totalPaidOut
        - activeBounties
        - completedBounties

    Transaction:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [escrow_lock, escrow_release, payout, deposit, refund]
        amount:
          type: number
        token:
          type: string
        fromWallet:
          type: string
          nullable: true
        toWallet:
          type: string
          nullable: true
        bountyId:
          type: integer
          nullable: true
        txSignature:
          type: string
          nullable: true
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - amount
        - token
        - description
        - createdAt

    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt

    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt

    CreateOpenaiConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title

    SendOpenaiMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    SendOpenaiVoiceMessageBody:
      type: object
      properties:
        audio:
          type: string
          description: Base64-encoded audio data
      required:
        - audio

    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - messages

    GenerateOpenaiImageBody:
      type: object
      properties:
        prompt:
          type: string
        size:
          type: string
          enum: ["1024x1024", "512x512", "256x256"]
      required:
        - prompt

    GenerateOpenaiImageResponse:
      type: object
      properties:
        b64_json:
          type: string
      required:
        - b64_json

    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error