# lib/api-zod/src/generated/types/bounty.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Frontier Road API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BountyStatus } from "./bountyStatus";

export interface Bounty {
  id: number;
  title: string;
  description: string;
  rewardAmount: number;
  rewardToken: string;
  status: BountyStatus;
  category: string;
  creatorWallet: string;
  claimerWallet?: string | null;
  proofOfWork?: string | null;
  escrowTxSignature?: string | null;
  completionTxSignature?: string | null;
  createdAt: Date;
  updatedAt: Date;
}