# lib/api-zod/src/generated/types/createOpenaiConversationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Frontier Road API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOpenaiConversationBody {
  title: string;
}