# lib/api-zod/src/generated/types/createResidentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Frontier Road API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateResidentBody {
  name: string;
  walletAddress?: string;
  avatar?: string;
  skills: string[];
  floor?: number;
  bio?: string;
}