# lib/api-zod/src/generated/types/resident.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Frontier Road API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ResidentStatus } from "./residentStatus";

export interface Resident {
  id: number;
  name: string;
  walletAddress?: string | null;
  avatar?: string | null;
  skills: string[];
  floor?: number | null;
  status: ResidentStatus;
  bio?: string | null;
  bountiesCompleted: number;
  bountiesCreated: number;
  totalEarned: number;
  createdAt: Date;
}