# lib/api-zod/src/generated/types/updateResidentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Frontier Road API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateResidentBodyStatus } from "./updateResidentBodyStatus";

export interface UpdateResidentBody {
  name?: string;
  walletAddress?: string;
  avatar?: string;
  skills?: string[];
  floor?: number;
  status?: UpdateResidentBodyStatus;
  bio?: string;
}